Added CanBeNative concern
