+++ packages/forms/src/Components/DateTimePicker.php
@@ -12,6 +12,7 @@ use Illuminate\View\ComponentAttributeBag;
 
 class DateTimePicker extends Field implements Contracts\HasAffixActions
 {
+    use Concerns\CanBeNative;
     use Concerns\CanBeReadOnly;
     use Concerns\HasAffixes;
     use Concerns\HasDatalistOptions;
@@ -35,8 +36,6 @@ class DateTimePicker extends Field implements Contracts\HasAffixActions
     protected ?int $firstDayOfWeek = null;
 
     protected string | Closure | null $format = null;
-
-    protected bool | Closure $isNative = true;
 
     protected bool | Closure $hasDate = true;
 
@@ -274,13 +273,6 @@ class DateTimePicker extends Field implements Contracts\HasAffixActions
         return $this;
     }
 
-    public function native(bool | Closure $condition = true): static
-    {
-        $this->isNative = $condition;
-
-        return $this;
-    }
-
     public function date(bool | Closure $condition = true): static
     {
         $this->hasDate = $condition;
@@ -465,11 +457,6 @@ class DateTimePicker extends Field implements Contracts\HasAffixActions
     public function shouldCloseOnDateSelection(): bool
     {
         return (bool) $this->evaluate($this->shouldCloseOnDateSelection);
-    }
-
-    public function isNative(): bool
-    {
-        return (bool) $this->evaluate($this->isNative);
     }
 
     public function getStep(): int | float | string | null
