+++ packages/forms/src/Components/Select.php
@@ -33,6 +33,7 @@ use function Filament\Support\generate_search_term_expression;
 class Select extends Field implements Contracts\CanDisableOptions, Contracts\HasAffixActions, Contracts\HasNestedRecursiveValidationRules
 {
     use Concerns\CanAllowHtml;
+    use Concerns\CanBeNative;
     use Concerns\CanBePreloaded;
     use Concerns\CanBeSearchable;
     use Concerns\CanDisableOptions;
@@ -82,8 +83,6 @@ class Select extends Field implements Contracts\CanDisableOptions, Contracts\Has
     protected ?Model $cachedSelectedRecord = null;
 
     protected bool | Closure $isMultiple = false;
-
-    protected bool | Closure $isNative = true;
 
     protected ?Closure $getOptionLabelUsing = null;
 
@@ -552,13 +551,6 @@ class Select extends Field implements Contracts\CanDisableOptions, Contracts\Has
         return $this;
     }
 
-    public function native(bool | Closure $condition = true): static
-    {
-        $this->isNative = $condition;
-
-        return $this;
-    }
-
     public function position(string | Closure | null $position): static
     {
         $this->position = $position;
@@ -692,11 +684,6 @@ class Select extends Field implements Contracts\CanDisableOptions, Contracts\Has
     public function isMultiple(): bool
     {
         return (bool) $this->evaluate($this->isMultiple);
-    }
-
-    public function isNative(): bool
-    {
-        return (bool) $this->evaluate($this->isNative);
     }
 
     public function isSearchable(): bool
